extract chunkid header
